1. Moved to common module Templates for all kinds of templates 2. Moved projects templates to Projects
