+++ lib/gitlab/client.rb
@@ -15,18 +15,14 @@ module Gitlab
     include Builds
     include Commits
     include Deployments
-    include DockerfileTemplates
     include Environments
     include Events
-    include GitignoreTemplates
-    include GitlabCiYmlTemplates
     include GroupMilestones
     include Groups
     include Issues
     include Jobs
     include Keys
     include Labels
-    include LicenseTemplates
     include MergeRequestApprovals
     include MergeRequests
     include Milestones
@@ -36,7 +32,6 @@ module Gitlab
     include PipelineTriggers
     include Pipelines
     include Projects
-    include ProjectTemplates
     include Repositories
     include RepositoryFiles
     include Runners
@@ -45,6 +40,7 @@ module Gitlab
     include Snippets
     include SystemHooks
     include Tags
+    include Templates
     include Todos
     include Users
     include Versions
