+++ spec/gitlab/client/projects_spec.rb
@@ -712,4 +712,43 @@ describe Gitlab::Client do
       expect(@file.markdown).to eq('[null](/uploads/f22e67e35e1bcb339058212c54bb8772/null)')
     end
   end
+
+  describe '.project_templates' do
+    before do
+      stub_get('/projects/3/templates/licenses', 'project_templates')
+      @project_templates = Gitlab.project_templates(3, 'licenses')
+    end
+
+    it 'gets the correct resource' do
+      expect(a_get('/projects/3/templates/licenses')).to have_been_made
+    end
+
+    it "returns a paginated response of project's templates" do
+      expect(@project_templates).to be_a Gitlab::PaginatedResponse
+    end
+  end
+
+  describe '.project_template' do
+    context 'when dockerfiles' do
+      before do
+        stub_get('/projects/3/templates/dockerfiles/dock', 'dockerfile_project_template')
+        @project_template = Gitlab.project_template(3, 'dockerfiles', 'dock')
+      end
+
+      it 'gets the correct resource' do
+        expect(a_get('/projects/3/templates/dockerfiles/dock')).to have_been_made
+      end
+    end
+
+    context 'when licenses' do
+      before do
+        stub_get('/projects/3/templates/licenses/mit', 'license_project_template')
+        @project_template = Gitlab.project_template(3, 'licenses', 'mit')
+      end
+
+      it 'gets the correct resource' do
+        expect(a_get('/projects/3/templates/licenses/mit')).to have_been_made
+      end
+    end
+  end
 end
